Moved things that Devan said were not necessary into TBD
