+++ src/TBD/ClientCommunicator.java
@@ -1,4 +1,4 @@
-package stockGenie;
+package TBD;
 
 import java.io.IOException;
 import java.io.ObjectInputStream;
