+++ src/stockGenie/StrategyOne.java
@@ -10,6 +10,8 @@ import com.tictactec.ta.lib.Core;
 import com.tictactec.ta.lib.MInteger;
 import com.tictactec.ta.lib.RetCode;
 
+import TBD.ClientGUI;
+
 
 public class StrategyOne extends Strategy {
